requirements.txt
email-validator
flask-login
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
requests
firebase-admin
sqlalchemy
werkzeug
python-dotenv
